% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cov.R
\name{add_cov}
\alias{add_cov}
\title{Simulate covariates with a multivariate normal distribution}
\usage{
add_cov(dt, covObj, seed)
}
\arguments{
\item{dt}{a \code{matrix} generated in \code{\link{set_n}}}

\item{covObj}{an object of class \code{.covClass} generated in \code{\link{set_cov}}}

\item{seed}{the seed of R‘s random number generator. Default is the first element of .Random.seed}
}
\value{
a \code{matrix} containing simulated covariates information
}
\description{
Simulate covariates with a multivariate normal distribution
}
\keyword{internal}
\keyword{method}
