% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_mcmc.R
\name{add_mcmc}
\alias{add_mcmc}
\title{Generating posterior samples from MCMC}
\usage{
add_mcmc(dt, priorObj, n.chains, n.adapt, n.burn, n.iter, seed)
}
\value{
A \code{list} containing hazard ratio and prior information
}
\description{
Generating posterior samples from MCMC
}
\keyword{internal}
\keyword{method}
