% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cov.R
\name{c,.covClass-method}
\alias{c,.covClass-method}
\title{Concatenate multiple \code{.covClasss} classes}
\usage{
\S4method{c}{.covClass}(x, ...)
}
\arguments{
\item{x}{A \code{.covClasss} class with covariate information generated in \code{\link{set_cov}}}

\item{...}{Other \code{.covClasss} classes with covariate information generated in \code{\link{set_cov}}}
}
\value{
A vector of \code{.covClasss} classes
}
\description{
Concatenate multiple \code{.covClasss} classes
}
\keyword{helper}
\keyword{method}
