% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_mcmc.R
\name{c,.priorClass-method}
\alias{c,.priorClass-method}
\title{Concatenate multiple \code{.priorClasss} class}
\usage{
\S4method{c}{.priorClass}(x, ...)
}
\arguments{
\item{x}{A \code{.priorClasss} class with prior distribution information generated in \code{\link{set_prior}}}

\item{...}{A \code{.priorClasss} class with prior distribution information generated in \code{\link{set_prior}}}
}
\value{
A vector of \code{.priorClasss} classes
}
\description{
Concatenate multiple \code{.priorClasss} class
}
\keyword{helper}
\keyword{method}
