% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_time.R
\docType{class}
\name{.clinClass-class}
\alias{.clinClass-class}
\alias{.clinClass}
\title{S4 Class for specifying parameters for enrollment time, drop-out pattern and analysis start time}
\description{
S4 Class for specifying parameters for enrollment time, drop-out pattern and analysis start time
}
\keyword{class}
