% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_time.R
\docType{class}
\name{.eventClass-class}
\alias{.eventClass-class}
\alias{.eventClass}
\title{S4 Class for setting parameters for time-to-events}
\description{
S4 Class for setting parameters for time-to-events
}
\keyword{class}
