% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_mcmc.R
\docType{class}
\name{.priorClass-class}
\alias{.priorClass-class}
\alias{.priorClass}
\title{S4 Class for specifying prior distributions and predictors for MCMC methods}
\description{
S4 Class for specifying prior distributions and predictors for MCMC methods
}
\keyword{class}
