% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_mcmc.R
\name{fix_col_names}
\alias{fix_col_names}
\title{Fix Column Names}
\usage{
fix_col_names(x, column_names)
}
\arguments{
\item{x}{a mcmc results object created by \code{\link[=add_mcmc]{add_mcmc()}}}

\item{column_names}{The names to change the beta columns to}
}
\description{
Utility function to make the mcmc column names more human friendly
}
