% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summary.R
\name{get_summary}
\alias{get_summary}
\title{Generate summary statistics of a simulation scenario}
\usage{
get_summary(dt)
}
\arguments{
\item{dt}{a \code{data.frame} containing summary statistics for the posterior samples from each simulation}
}
\value{
a \code{data.frame} containing the mean and sd of posterior HR between treatment and control arm, the posterior mean and sd of HR between internal control and external control arm, reject rate, variance, bias and mse of the simulation set
}
\description{
Generate summary statistics of a simulation scenario
}
\keyword{method}
