% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_psborrow_dev}
\alias{is_psborrow_dev}
\title{Check if user is in psborrow development environment}
\usage{
is_psborrow_dev()
}
\value{
TRUE/FALSE flag (TRUE = in development environment)
}
\description{
Simple function which leverages the DESCRIPTION file
to check if the user is in a development environment
for psborrow.
}
