% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_cov.R
\name{match_cov}
\alias{match_cov}
\title{Match}
\usage{
match_cov(dt, match)
}
\arguments{
\item{dt}{a list of \code{matrix}}

\item{match}{A vector of covariates name to match on}
}
\description{
Match
}
\keyword{constructor}
