% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summary.R
\name{plot_bias}
\alias{plot_bias}
\title{Plot bias}
\usage{
plot_bias(dt, HR = 1, driftHR = 1, pred = "none")
}
\arguments{
\item{dt}{a \code{data.frame} containing summary statistics for the posterior samples from each simulation generated with \code{\link{get_summary}}}

\item{HR}{pre-specified HR between treatment and control arm in the internal trial}

\item{driftHR}{pre-specified HR between external control arm and internal control arm}

\item{pred}{predictors to use when fitting exponential distribution in MCMC}
}
\value{
a \code{ggplot} which is a bar plot containing bias implications corresponding to each prior, the pre-specified HR between internal treatment and control arms is 1
}
\description{
Plot bias
}
\keyword{method}
