% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summary.R
\name{plot_hr}
\alias{plot_hr}
\title{Plot posterior hazard ratio between treatment and control}
\usage{
plot_hr(dt, HR = 0.67, driftHR = 1, pred = "none")
}
\arguments{
\item{dt}{a \code{data.frame} containing summary statistics for the posterior samples from each simulation generated with \code{\link{get_summary}}}

\item{HR}{pre-specified HR between treatment and control arm in the internal trial}

\item{driftHR}{pre-specified HR between external control arm and internal control arm}

\item{pred}{predictors to use when fitting exponential distribution in MCMC}
}
\value{
a \code{ggplot} which is a point plot containing posterior hazard ratio between treatment and control arms corresponding to each prior
}
\description{
Plot posterior hazard ratio between treatment and control
}
\keyword{method}
