% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ps_message}
\alias{ps_message}
\title{Conditional Message}
\usage{
ps_message(...)
}
\arguments{
\item{...}{Values passed onto \code{\link[=message]{message()}}}
}
\description{
Simple wrapper function around \code{\link[=message]{message()}} that will supress
printing messages if the option \code{psborrow.quiet} is set to \code{TRUE}
i.e.\preformatted{options("psborrow.quiet" = TRUE)
}
}
