% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summary.R
\name{rej_est}
\alias{rej_est}
\title{Generate summary statistics for the MCMC chains}
\usage{
rej_est(samples)
}
\arguments{
\item{samples}{an object of class \code{mcmc.list}}
}
\value{
a vector containing the mean, median, sd, reject rate for the MCMC chains
}
\description{
Generate summary statistics for the MCMC chains
}
\keyword{method}
