% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpwexp.R
\name{rpwexp}
\alias{rpwexp}
\title{The Piecewise Exponential Distribution}
\usage{
rpwexp(n, rate = 1, intervals = NULL, cumulative = FALSE)
}
\arguments{
\item{n}{Number of observations.
If length(n) > 1, the length is taken to be the number required.}

\item{rate}{Vector containing exponential failure rates in intervals described by
\code{intervals}}

\item{intervals}{Vector containing positive values indicating interval lengths where
the exponential rates provided in \code{rate} apply. Note that the length of
\code{intervals} is 1 less than that of \code{rate} and that the final value rate
in \code{rate} applies after time \code{sum(intervals)}.}

\item{cumulative}{\code{FALSE} (the default) generates \code{n} independent,
identically distributed piecewise exponential failure rates according to the distribution
specified by \code{intervals} and \code{rate}. \code{TRUE} generates independent
inter-arrival times with the rates of arrival in each interval specified by
\code{intervals} determined by \code{rate}.}
}
\description{
The piecewise exponential distribution allows a simple method to specify a distribtuion
where the hazard rate changes over time. It is likely to be useful for conditions where
failure rates change, but also for simulations where there may be a delayed treatment
effect or a treatment effect that that is otherwise changing (e.g., decreasing) over time.
\code{rpwexp()} is to support simulation both the Lachin and Foulkes (1986) sample size
method for (fixed trial duration) as well as the Kim and Tsiatis(1990) method
(fixed enrollment rates and either fixed enrollment duration or fixed minimum follow-up);
see \code{\link[gsDesign:nSurv]{gsDesign}}.
}
\details{
Using the \code{cumulative=TRUE} option, enrollment times that piecewise constant over
time can be generated.
}
