% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mcmc.R
\name{run_mcmc}
\alias{run_mcmc}
\title{Run MCMC for multiple scenarios with provided data}
\usage{
run_mcmc(dt, priorObj, n.chains, n.adapt, n.burn, n.iter, seed, path)
}
\arguments{
\item{dt}{a list of \code{matrix} containing simulated time-to-events information}

\item{priorObj}{an object of class \code{.priorClass} generated in \code{\link{set_prior}}}

\item{n.chains}{number of parallel chains for the model}

\item{n.adapt}{number of iterations for adaptation}

\item{n.burn}{number of iterations discarded as burn-in}

\item{n.iter}{number of iterations to monitor}

\item{seed}{the seed of random number generator. Default is the first element of .Random.seed}

\item{path}{file name for saving the output including folder path}
}
\value{
a \code{data.frame} containing summary statistics of the posterior distribution for each simulation
}
\description{
Run MCMC for multiple scenarios with provided data
}
\examples{
# examples in vignette

}
\keyword{simulator}
