% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_png}
\alias{save_png}
\title{Save .png file (for snapshot testing purposes)}
\usage{
save_png(plot, width = 1200, height = 800)
}
\arguments{
\item{plot}{plot to be printed}

\item{width}{in pixels}

\item{height}{in pixels}
}
\value{
temp path to the image
}
\description{
Save .png file (for snapshot testing purposes)
}
