% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cov.R
\name{set_n}
\alias{set_n}
\title{Simulate external trial indicator and treatment arm indicator}
\usage{
set_n(ssC, ssE, ssExt)
}
\arguments{
\item{ssC}{Number of observations in the internal control arm. Default is 100}

\item{ssE}{Number of observations in the internal experiment arm. Default is the same number
of observations as \code{ssC}}

\item{ssExt}{Number of observations in the external control arm. Default is the same number
of observations as \code{ssC}}
}
\value{
A \code{matrix} containing external trial indicator and treatment indicator
}
\description{
This function conducts validity check and generates a matrix with two binary variables indicating
\enumerate{
\item if the observation belongs to the external trial
\item if the observation belongs to the treatment arm.
}
}
\keyword{constructor}
