% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu_time.R
\name{simu_time}
\alias{simu_time}
\alias{simu_time,matrix,.eventClass-method}
\title{Simulate time-to-events for multiple scenarios}
\usage{
simu_time(dt, eventObj, clinInt, clinExt, seed, path)
}
\arguments{
\item{dt}{a list of \code{matrix} generated in \code{\link{simu_cov}} containing simulated covariates information}

\item{eventObj}{an object of class \code{.eventClass} generated in \code{\link{set_event}} including event information}

\item{clinInt}{an object of class \code{.clinClass} generated in \code{\link{set_clin}} including internal trial information}

\item{clinExt}{an object of class \code{.clinClass} generated in \code{\link{set_clin}} including external trial information}

\item{seed}{the seed of random number generator. Default is 47}

\item{path}{file name for saving the output including folder path}
}
\value{
a list of \code{matrix} containing simulated time-to-events information
}
\description{
Simulate time-to-events for multiple scenarios
}
\keyword{simulator}
