% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_mcmc.R
\name{valid_mcmc}
\alias{valid_mcmc}
\title{validate the operational parameters for MCMC method}
\usage{
valid_mcmc(n.chains, n.adapt, n.burn, n.iter)
}
\description{
validate the operational parameters for MCMC method
}
\keyword{internal}
\keyword{method}
