% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cov.R
\name{add_cov}
\alias{add_cov}
\title{Simulate covariates with a multivariate normal distribution}
\usage{
add_cov(dt, covObj, seed)
}
\value{
a \code{matrix} containing simulated covariates information
}
\description{
Simulate covariates with a multivariate normal distribution
}
\keyword{internal}
\keyword{method}
