% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_mcmc.R
\name{add_mcmc}
\alias{add_mcmc}
\title{Generating posterior samples from MCMC}
\usage{
add_mcmc(dt, priorObj, n.chains, n.adapt, n.burn, n.iter, seed)
}
\arguments{
\item{dt}{a list of \code{matrix} containing simulated time-to-events information}

\item{priorObj}{an object of class \code{.priorClass} generated in \code{\link{set_prior}}}

\item{n.chains}{number of parallel chains for the model}

\item{n.adapt}{number of iterations for adaptation}

\item{n.burn}{number of iterations discarded as burn-in}

\item{n.iter}{number of iterations to monitor}

\item{seed}{the seed of random number generator. Default is the first element of .Random.seed}
}
\value{
A \code{list} containing hazard ratio and prior information
}
\description{
Generating posterior samples from MCMC
}
\keyword{internal}
\keyword{method}
