% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cov.R
\name{c,.covClass-method}
\alias{c,.covClass-method}
\title{Concatenate multiple \code{.covClasss} classes}
\usage{
\S4method{c}{.covClass}(x, ...)
}
\arguments{
\item{x}{A \code{.covClasss} class with covariate information generated in \code{\link{set_cov}}}

\item{...}{Other \code{.covClasss} classes with covariate information generated in \code{\link{set_cov}}}
}
\value{
A vector of \code{.covClasss} classes
}
\description{
Concatenate multiple \code{.covClasss} classes
}
\examples{
# combine two sets of covariates
covset1 = set_cov(n_cat = 2, n_cont = 0, mu_int = 0, mu_ext = 0, var = 1)
covset2 = set_cov(n_cat = 0, n_cont = 1, mu_int = 62, mu_ext = 65, var = 11)
cov_list = c(covset1, covset2)



}
\keyword{helper}
\keyword{method}
