% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summary.R
\name{plot_type1error}
\alias{plot_type1error}
\title{Plot type 1 error}
\usage{
plot_type1error(dt, driftHR = 1, pred = "none")
}
\arguments{
\item{dt}{a \code{data.frame} containing summary statistics for the posterior samples from each simulation generated with \code{\link{get_summary}}}

\item{driftHR}{pre-specified HR between external control arm and internal control arm}

\item{pred}{predictors to use when fitting exponential distribution in MCMC}
}
\value{
a \code{ggplot} which is a bar plot containing type 1 error implications corresponding to each prior, the pre-specified HR between internal treatment and control arms is 1
}
\description{
Plot type 1 error
}
\keyword{method}
