% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu_cov.R
\name{simu_cov}
\alias{simu_cov}
\alias{simu_cov,matrix-method}
\title{Simulate covariates}
\usage{
simu_cov(ssObj, covObj, driftHR, HR, nsim, seed, path)
}
\arguments{
\item{ssObj}{an object of class \code{.covClass} generated in \code{\link{set_n}}}

\item{covObj}{an object of class \code{.covClass} generated in \code{\link{set_cov}}}

\item{driftHR}{hazard ratio of external control and internal control arms}

\item{HR}{a list of hazard ratio of treatment and control arms}

\item{nsim}{number of simulation. Default is 5}

\item{seed}{the seed of random number generator. Default is 47}

\item{path}{file name for saving the output including folder path}
}
\value{
a list of \code{matrix} containing simulated covariates information
}
\description{
This function generates continuous and binary covariates through simulating from a multivariate normal distribution. Outcomes are further converted to binary variables using quantiles of the normal distribution calculated from the probability provided. Then the covariates are added to the external trial and treatment arm indicators.
}
\keyword{simulator}
