% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_mcmc.R
\name{valid_mcmc}
\alias{valid_mcmc}
\title{validate the operational parameters for MCMC method}
\usage{
valid_mcmc(n.chains, n.adapt, n.burn, n.iter)
}
\arguments{
\item{n.chains}{number of parallel chains for the model}

\item{n.adapt}{number of iterations for adaptation}

\item{n.burn}{number of iterations discarded as burn-in}

\item{n.iter}{number of iterations to monitor}
}
\value{
A \code{list} containing numbers of parallel chains and iterations
}
\description{
validate the operational parameters for MCMC method
}
\keyword{internal}
\keyword{method}
